rewrite code after comments
