+++ package.json
@@ -6,6 +6,7 @@
   "author": "Mate Academy",
   "license": "GPL-3.0",
   "dependencies": {
+    "-": "^0.0.1",
     "@cypress/react": "^5.12.4",
     "@fortawesome/fontawesome-free": "^6.1.2",
     "bulma": "^0.9.4",
@@ -26,9 +27,12 @@
     "@types/node": "^17.0.23",
     "@types/react": "^17.0.43",
     "@types/react-dom": "^17.0.14",
+    "@typescript-eslint/eslint-plugin": "^5.40.1",
+    "@typescript-eslint/parser": "^5.40.1",
     "cypress": "^9.5.3",
     "eslint": "^7.32.0",
     "eslint-plugin-cypress": "^2.11.2",
+    "eslint-plugin-react": "^7.31.10",
     "gh-pages": "^3.2.3",
     "mochawesome": "^7.1.3",
     "mochawesome-merge": "^4.2.0",
