+++ src/App.tsx
@@ -6,14 +6,10 @@ import {
   Route,
   Routes,
 } from 'react-router-dom';
-import { PageNavLink } from './components/PageNavLink';
 import { TabsPage } from './components/TabsPage';
-
-const tabs = [
-  { id: 'tab-1', title: 'Tab 1', content: 'Some text 1' },
-  { id: 'tab-2', title: 'Tab 2', content: 'Some text 2' },
-  { id: 'tab-3', title: 'Tab 3', content: 'Some text 3' },
-];
+import { HomePage } from './components/HomePage';
+import { PageNotFound } from './components/PageNotFound';
+import { PageNav } from './components/PageNav';
 
 export const App = () => (
   <>
@@ -23,16 +19,7 @@ export const App = () => (
     >
       <div className="container">
         <div className="navbar-brand">
-          <PageNavLink
-            to="/"
-            text="Home page"
-            end
-          />
-
-          <PageNavLink
-            to="/tabs"
-            text="Tabs page"
-          />
+        < PageNav />
         </div>
       </div>
     </nav>
@@ -42,7 +29,7 @@ export const App = () => (
         <Routes>
           <Route
             path="/"
-            element={<h1 className="title">Home Page</h1>}
+            element={<HomePage/>}
           >
             <Route
               path="home"
@@ -51,12 +38,12 @@ export const App = () => (
           </Route>
 
           <Route path="tabs">
-            <Route index element={<TabsPage tabs={tabs} />} />
-            <Route path=":tabId" element={<TabsPage tabs={tabs} />} />
+            <Route index element={<TabsPage />} />
+            <Route path=":tabId" element={<TabsPage />} />
           </Route>
           <Route
             path="*"
-            element={<h1 className="title">Page not found</h1>}
+            element={< PageNotFound />}
           />
         </Routes>
       </div>
