+++ src/components/HomePage.tsx
@@ -1,0 +1,3 @@
+export const HomePage=()=>(
+  <h1 className="title">Home Page</h1>
+)
