+++ src/components/PageNav.tsx
@@ -1,0 +1,15 @@
+import { PageNavLink } from "./PageNavLink";
+
+export const PageNav = () => (
+  <>
+    <PageNavLink
+      to="/"
+      text="Home page"
+      end
+    />
+    <PageNavLink
+      to="/tabs"
+      text="Tabs page"
+    />
+  </>
+);
